[AT-SPI] Make DevelControl::GetAccessibilityRelations() equivalent to Accessible::GetRelationSet()
The previous std::vector<std::vector<Address>> return type, in which the
RelationType enumeration value served as an index to the outer vector,
can be changed to std::vector<Relation>, which is more compact (there
are very few relations per control in practice). Also, the relation
target type is now Accessible*, which is more useful than Address.

Change-Id: I9f7b86d7415e1711b8a81cd66e271ff9cdffd5c2