Fix uint16 max issue in TextLabel
If the size of TextLabel is bigger than uint16 max value,
there is a problem that assert occurs in Texture::New.

We can avoid the assert by setting the VisualTransform size to uint16 max,
but this also doesn't guaranteed rendering behavior.

To guaranteed the behavior, VisualTransform size is limitted to maxTextureSize.

Change-Id: I68523c89aecc71a9158e59948bca037d5c73c95a
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>