Fix crash issue at VisualBase (DoAction + AnimateTo)
Case 1 :

When we try to change decorated properties by DoAction and UPDATE_PROPERTY,
It can be crashed due to the VisualRenderer is not DecoratedVisualRenderer.
This patch make we update properties only for valid visual types.

Case 2 :

When we try to animate with DevelVisual's animatable property, It become crashed.
This cause we miss some logic of found registered property.
Now, we can run arc-visual.example

Change-Id: Icee700608277a4133426bb39a4c30b92ab8acc8f
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>