[Tizen] Erase observer at TexturManager remove
First, load one image from imageView.
And while the first loaded image is removed from the scene, there is a problem if another image view loads with the same image URL.
This is because there are cases where a callback is received from an invalid imageView by a previous observer.
To prevent this problem, change the observer to also delete the texture when deleting it.

Change-Id: Ied8235c35ff9772b36f55de3c3b5b21fbc4bce5c
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>