Fix cursor position update issue
When a text controller has focus and loses keyboard focus,
the controller is in the INACTIVE state.
But in this case, when the cursor position is updated from a key event,
there is an issue that the position is not updated due to the wrong state.

This patch changes the state of controller to EDITING in the issue case.

Change-Id: Idea21a5486a7c076a59ad5dcd74c5b3343c1e2fe
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>