Fix GetTextBoundingRectangle size issue
TextGeometry calculates the size and position of the actual text with glyphAdvance.
This means that sometimes the x position will be negative,
or the text's line width may be greater than the control's width.

While this is an accurate value,
it is not suitable for calculating the minimum bounding rect of text.
Because the text appears to be outside the control.

This patch adjusts the return value of GetTextBoundingRectangle()
so that it does not exceed the bounds of the text control.

Change-Id: Ie80ac1a8b5cec5f8c21013744645254381c78fb6
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>