Adjust text fit condition
If the calculated text size and layout size are the same in the text fit case,
there is an issue of ellisis processing due to the influence of some glyph's advance.
So in most cases the layout needs a bit larger size.
This patch adds a condition to prevent this issue.

Change-Id: I4592c8741d19cce6d33474288d55957a3944149d
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>