[Tizen] Fix TextUpdateInfo::Clear() issue
The Clear method sets (uint)mCharacterIndex to -1,
so it makes a maximum uint value.
This causes many problems, but since it was implemented,
all index logic is related to it, so it is difficult to fix.

This patch avoids this issue by backup/restore of mCharacterIndex like TextEditor.
But we should consider changing mCharacterIndex to an integer.

Change-Id: Ida2c2af83438e71f6f05713a7de927f7b15aa7bd
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>