[Tizen] Be a bit smarter with SetFocusedActorPosition
The previous calculation method was to always use the center point of a given
actor, regardless of the position of the touch event that caused it to be
focused. In the tricky scenario where a smaller actor is overlaid on top of a
bigger actor such that the smaller actor covers the center of the bigger actor,
a double-tap-and-hold gesture meant to activate the bigger actor would activate
the smaller actor instead.

The calculation method introduced by this patch is a little bit better, but
still not ideal. However, the HighlightedObjectInfo D-Bus method has only two
parameters, the coordinates, taken from GetFocusedActorPosition, which are then
forwarded back into the app as a touch event, so there isn't much more that
could be done in DALi to address this issue.

Change-Id: I03f291f3dece2bd2c10fedd7ce22f8791230b6f4