Fix space font validation condition
when character is 0x20, there are various cases of font validation.
but if it is a default font or cached font, or
if a glyph exists in the current font when it is contiguous with the prev char,
it is natural to load current font.
this is same behaviour as before.. and the process of finding fonts can slightly reduced.

Change-Id: I017db2ec38e86642a83a44739a0d2c7af072e14a