[Tizen] Load bvh and facial animations from buffer
This is a combination of 2 commits.

Make sure that global variables are initialized lazily in scene3d.

Global variables are initialized when scene3d so is loaded firstly.
This would allocate more memory in some cases.
This patch is to make the variables be initialized lazily.

Load bvh and facial animations from buffer

Let we allow to load bvh and facial animation from raw buffer stream.

Change-Id: I679539ce10e631101571dab95615404c41b9b2f3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>