Blendshapes fix (still missing tests)
When using sparse accessors, the gltf loader loads data correctly
but the blendshapes generator (texture and data) always assumes that
morph targets contain all the vertices which isn't a case.

Fix makes sure that for sparse access only indexed vertices are
affected by the blendshape. It still generate full vertex data
(for whole mesh) so it's compatible with current shader.

Change-Id: I1f36c1d9d24dec0eb194d0aaf1b93488b61b1d58
Signed-off-by: Adam Bialogonski <adam.b@samsung.com>