[Tizen] Fix unload svg file error.
When we create Svg Visual -> SceneOn & SceneOff very fast during all rasterize
threads are fully busy,
It will remove SvgLoadingTask who is in mResterizeTasks.

After we re-sceneon that visual again, it will be broken :
mLoadingFail is not true, but we didn't load file.

This patch make we don't remove SvgLoadingTask when we call RemoveTasks

Change-Id: I40582ebb2512e56e96eed64caff4b1070b4bd9c2
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>