Fix GetHeightForWidth for text controller
In the case of text label,
the height of the layout is used when returning the cached value of GetHeightForWidth.

But, this value is the height of the layout that has been Ellipsis processed.
It's not the height of the entire Text.

So, the intended height cannot be obtained.

This patch simply stores calculated values.

Change-Id: Ia5d06bdac39556776749de3e3078a3a7e230c36c
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>