Add GetHeightForWidth for text visual model
Use visual model's GetHeightForWidth instead of GetLayoutSize.

In the case of text label,
the height of the layout is used when returning the cached value of GetHeightForWidth.

But, this value is the height of the layout that has been Ellipsis processed.
It's not the height of the entire Text.

So, the intended height cannot be obtained.

This patch simply stores calculated values.

Change-Id: I0494fb476964fb1207d35e3c8e6eb25ece972025
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>