(AnimatedVector) Cache lottie file's layer / marker info
GetLayerInfo and GetMarkerInfo can be called usually if we use NUI system.
(Since we call Visual::Base::CreatePropertyMap() every times we change some
properties.)

But the LayerInfo / MarkerInfo never be changed if the file is fixed.
So let we cache that information, and use cached value.

It can also reduce mutex block at plugin system internally.

Change-Id: I767bd581dc5e2a41715263d03c1ace15d2476df3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>