Let Anti-Alias consider scale factor
Some days ago, we assume that the view which is smaller than 100x100 doesn't use
the range of anti-alias as 1.0.

(See https://review.tizen.org/gerrit/c/platform/core/uifw/dali-toolkit/+/279801)

But it make some visual defect when we use 56x56 size of view and scale it up.

Basically, the problem of these issues comes due to visuald don't know the
scale value.
So let we get actor's scale value as default shader property,
and then use it as anti-alias calculation parameter.

Change-Id: I64b32049919cd79c9dda7fcf7c9c256e5b5a6078
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>