Append 1 pixel if corner radius is full-circle
Their was some visual quality issue occured when we render perfect circle.

The reason of issue was, the fragments for soft-anti-alias area was required.

For resolve this issue greedly, let we just append vertex position margin,
and render it applied as that margin value.

TODO : Currently, we only consider fully-circle case. Let we need to calculate
is there any similar defect need to be resolved for normal rect cases.

Change-Id: Ib114c29b08806e21c3830b1dc65a8098b34ef060
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>