Remove cubemap cache temperary + Fix bug when Uploaded PixelData resue after Texture GC
If we create NewPixelDataWithReleaseAfterUpload and Upload to texture,
and if that texture reference count become 1 and GC,
the key of PixelData is not valid anymore.

To avoid this cases, Let we collect all PixelData container if some texture removed.
And also, since PixelData could be key of Texture cache, we should check the
reference count is less than 2.

+

Since current cubemap cache not works well reasonable, just remove it.
We will reopen it after we implement PixelDataList cache system again.

+

Print more log information if we need.

Change-Id: I6693ceab9b03cc20e1c6a42282ef39af38344e27
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>