[Tizen] Fix the 3D model rendering issue after the KHR_texture_transform etension support
Revert "[Tizen] Fix bug that IBLFactor / Roughness not works well when we change material"

This reverts commit 519b293f4452ddb4177af714e841c2b68a7335e5.

Fix the 3D model rendering issue after the KHR_texture_transform extension support

Fix bug that IBLFactor / Roughness not works well when we change material

There was some bug when we change Material info, uMaxLOD and uIblIntensity does not applied what primitive has.

This is because Material's SetUniform API overwrite the value of those uniform.

This patch make we ensure the order of uniform register, so let we use correct LOD and IBL intensity.

Change-Id: Id8e11744a9767402aa0e36e112093b79fb9b8f3f
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>