(Scene3D) Optimize AnimatedProperty's KeyFrame values after load complete
Let we reduce the total key frame of MotionValue what we parse.

Since both json and bvh store the data by framerate.
So, there might be useless values are included.

If the number of keyframe is large, it might give some overhead to core.
So let we reduce useless keyframe for MotionValue and Model's internal animation.

Change-Id: I04b86b0341125cef1f60d3da68150c9367655169
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>