[Tizen] Fix editable text font size update issue
There is an issue that the font runs of the editable text area are separated
when the font size is updated.

If the editable text is an active state and the font size is updated,
the input font size should also be updated.

And due to the feature that changes the selected area's font size when the input font size is updated,
unintended font run separation and font size change occur if there is selected text when updating the font size.

This patch fixed so that the selection-related logic is not executed
when SetInputFontPointSize is called due to an update of the default font size.

Change-Id: I1a5f5190d00a25291a8cafc215f5bf0df2cd71f7
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>