(Vector) Let we send finished signal only for matched sent play/stop from visual
It is possible that AnimationFinished trigger comes "after" event thread Play action comes.

If then, user's Play action be skipped.

To avoid this case, let we give the id when play state changed to Play / Pause / Stop.
And send finished signal only if the id is matched what visual know.

Change-Id: I03409e65d2f6b947940ea5591959b106ff8de2bd
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>