Make depth-index as 'AUTO' instead of CONTENT change as default, and clamp depth-index
Let we set depth index as auto if user doesn't set visual's depth index.

Previously we use same behavior if user doesn't set, or user set depth index as CONTENT.
So if user set as 0, it might not works well.

To avoid this confuse logic, let we seperate whether user didn't set the property or not.

Furthermore, let we hard-clamp the range of DepthIndex for VisualBase.

The case of AUTO_INDEX is only for special cases, and otherwise, let we use normal values
what user set.

---

We allow to use RegisterVisual use depth-index as AUTO_INDEX. If this case, it will works
like what we don't set the depth index.

Change-Id: If19b0757b5e27e0a94a37fc8bc0f57c6824d0b16
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>