Less call mAnimatedImageLoading.GetImageCount() who might load file synchronously
Since AnimatedImageLoading might not load imagefile if we use
cached texture.

In this case, If we try to call 'GetImageCount()' or similar API, it will
try to load file synchronously, what user never notify.

For example, when we call DoCreatePropertyMap, it will call that API
what we can use some other cached value.

Change-Id: I77d30b985efd86d09663dfd51e885ba6227adbc9
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>