Make AnimatedImageVisual use single fixed image cache if it is not gif/webp + Fix several bugs at fixed image cache
It is possible that mImageCache is NULL if url suffix is not gif or webp.
In this case, mImageCache become null, so we should show broken image.

But also, we 'might' need to show non-animatable image even if we use
non-animatable image (like jpg), and set it to animated image visual forcibly.

To resolve general cases, let we make AnimatedImageVisual with non-animatable format image
just use image sequence with length 1.

===

Also, there was several bugs when we use fixed image cache, with cached texture manager image.
Before, we don't consider full-scenario when LoadComplete callback comes
during TextureManager.Load.

Change-Id: I173020e42d6447ff43e56e19f25ea8e06c7bbfc1
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>