Move image/ text / npatch relative class under each visual folder
For example, npatch-loader only be used for npatch-helper and npatch-visual,
we don't need to keep it under 'visuals' folder directly.

Similarly, text-visual-shader-factory only be used at text-visual.

image-visual-shader used several visuals, but all of them are used for image,
so move them  under image folder is not a bad job.

To make clear ownership of logic, let we move files into each folder.

Change-Id: Iaa1223c687425041a9a62e11792b0940ea8429cb
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>