Allow to set render effect for off-scene control + Auto deactivate if invisible
Let we don't activate the render effect if owner control is not on scene.

And also, auto deactivate if view is invisible + auto activate
if view is visible again.

And Deactivate & Activate is scensitive as changeness of owner control,
Let we Deactivate first before we change the owner control

TODO : Currently we don't consider the SceneHolder's visibility.
It will be fixed after InheritVisibilityChanged signal consider Scene's visibility.

Change-Id: If1f3cdc176a28135b9f7a3184e736b375357a1b3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>