Do not print whole image buffer
Since oss << uint8_t* try to print it as string, not a pointer,
it might print some strange things, and sometimes it might give crash.

To avoid this security issue, let we don't print uint8_t informations

And also, their was no way to seperate normal buffer, and embedded buffer.
So let we seperate keyword of it.

Change-Id: Ide223482b3cc86315488afab3c888db72146555b
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>