Ensure strncmp compare include null-terminate char
If we use n-paramater of strncmp as strlen, it will check only
Prefix of string.

For example,
char a[] = "asdf";
char b[] = "as";

then, strncmp(a, b, strlen(b)) will return 0, even if a and b is not equal.

To avoid this kind of problem, let we ensurely check end of comparitor is
delim or not.

Change-Id: I34b10a014cb06721af2cc92df06892bfb20bf6f3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>