Make color visual always use premulitplied alpha
Until now, ColorVisual::Property::MIX_COLOR and Visual::Property::OPACITY is
independence properties. So we cannot assume that mixColor uniform was
premultiplied or not. (This is why we cannot enable premultiplied to color visual)

But now we can allow to premultiply option to color visual.

More over, enable/disable premultiply option is meanless for color visual.
So let we use always pre-multiplied alpha option enabled - for performance.

Change-Id: I945e5362207c3b2f94c814d74a00988257fc7f52
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>