Ask StyleManager::Configuration only we need at TextController
We don't need to know the mShouldClearFocusOnEscape value for TextLabel.
And also, if input key is never be Escape, we also don't need it.

But that value required to ask singletone class, StyleManager,
and StyleManager might try to parse json file synchronously.

And also, for every TextVisual, we try to get Configure and find some value
by string, what we don't need to know.

It might make overhead for TextLabel creation.

To reduce the overhead, let we get the value only if we need.

Change-Id: I744cff4b05f8ed83d7924c4cbff3ed8f209a5571
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>