Fix mis-using of RegisterProperty + Sync AnimatedImageVisual logic with ImageVisual
Their was several mis-ussing cases for RegisterProperty.

Since the first value of RegisterProperty is the key of custom-property.
But some codes works like "Set or Register".

It will return invalid values when we try to get some item
by INVALID_KEY.

+

Some codes are mis-implements at AnimatedImageVisual or AnimatedVectorVisual, SvgVisual.
Let we follow them as latest code scenario, and add some UTC for it.

Change-Id: I2a5b23a1cc321b2fc1fdd880ba85b8859451c153
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>