Return empty map and remove builder at Configuration getter if builder load failed
It is possible that default theme builder load failed.
Most of cases consider it, but StyleManager::GetConfiguration() didn't care that case.

To ensure it, let we get LoadJSON result, and reset builder if load failed.
And return empty Property::Map.

*****************************************************************************************
NOTE 20250326:

utc-Dali-StyleManager uses toolkit-text-abstraction.
The segmentation fault seems to be caused by not finding the symbol.
I added the missing APIs to the fake font client and checked that the problem is solved.

+ Excluded unrelated files from the commit.
*****************************************************************************************

Change-Id: Icb40e69c19c48baa4c6b4350e6a495343e2cb47a
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>