[Tizen] Fix coverity issue : Dali::Vector<>.Size could be zero
Let's avoid size_t type overflow issue during binary search.

It could be broken if mMarks.Size() was 0u.
Also, if marks property value was not a float, also could be break (unlimited loop)

To avoid these kind of issue, let we change the logic of binary search
that is safe enough for overflow case.

Change-Id: I5b27daf0915c64e9192d4821b7d11e7e5a45f356
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>