[Tizen] Prevent atlas block corruption issue when renderer is reset
If UpdateTextType is NONE_UPDATED when Renderer is reset, RenderText does not perform rendering.
This causes several problems with atlas operation.

When the renderer is reset, all caches used by the text field are removed from AtlasManager.
At this time, if a cache of an ID that was previously used (now removed) by another editable text is created and used,
the atlas block of the existing text field will be damaged.

At least until the cache structure is improved, render must be performed when the renderer is reset.

Change-Id: Ic5f939e81920ab59e676ee50f450c6ffbb1259e3
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>