[Tizen] (TextAtlas) Let Text relative geometry use uint32_t indices
Due to the data format is uint16_t type, we cannot render well
if the length of text is bigger than 16384.
If we add underline or strokethrough, the number of text limitation
also be reduced.

To fix this issue, let we make indices what TextAtlas geometry use
uint32_t type, instead uint16_t type.

Since we support uint32_t* indices now a day, it will be okay.

Change-Id: I887d78bb5f7123580fc7d6d5b68e35a7d2fdd954
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>