[Tizen] Fix bidirectional line info index issue
In the RTL case, when text is inserted or removed in the middle of a bidi paragraph,
the character index of line info is not updated,
so there is a problem that the line info is damaged in reorders.

This patch updates the index of line info based on the bidi paragraph info updated in SetBidirectionalInfo.

Change-Id: Ifc1ddfa5f15ce89be967ce57241a093030be323d
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>