Reduce Visual::Base handle creation time + Implement SetTransformSize with empty transform map
If we are try to use internal side visual operation required,
we can use BaseObjectPtr directly, instead of wrapping the handle.

Since atomic reference count spend time if we call it frequencly,
let we use some codes use Internal::Visual::Base directly.

+

Most of case, if fitting mode not applied, we only need the size of control
and don't care the transform map.

But the transform map is mendentary input. So their are several
process works to create empty map and destroy it again.

To avoid this logic, let we make Visual::Base::SetControlSize()
and make we re-use the empty property map here.

Change-Id: I379cc226cdf2af9fea12bc0e2c9ea17d62a696c5
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>