Fix minor render scale issues
Integerizing the min line size makes the result more unnatural,
so I change it to a simple calculation.

And as I kept adding complex formulas to compensate for the real error of layout size,
the final result became the same as width / scale.
I will keep this state until I find a better way.

Change-Id: Ica8e456567db14d7a0ce2848521b35a85494edc6
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>