Fix emoji presentation error when using incorrect vs16
The main problem is that there is no variation selector in the samsung color emoji font.

However, the variation selector is not a glyph that is rendered,
so it should work regardless of whether the glyph exists in the font.

This patch finds the color font regardless of whether the vs glyph exists in the problem.

Related:
https://review.tizen.org/gerrit/c/platform/core/uifw/dali-adaptor/+/324809

Change-Id: I4e29db1b15878c5f3b373ecd0b0551e89a15c120
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>