(Vulkan) Updating default pbr shader
Material shader constants are translated in the shader as follows:

Any instance of "#ifdef FEATURE" are either ignored if not set, or,
translated into "#if 1" if set.

The pbr shader also defines"MORPH", regardless. This forces the blend
shape texture to be defined; however; if the morph shader features are not
defined (e.g. the model doesn't have blendshapes), then GLES aggressively
optimizes the blend shape texture out.

However, in Vulkan, we don't optimize enough, so the texture is still
defined, and the bind order is off by 1.

To fix this problem, we remove the blanket #define MORPH line, and instead
define it if any of the shader constants are translated into "#if 1".

This also highlighted another difference between GLSL and SPIRV - the latter
defines gl_VertexIndex, whereas GLES uses gl_VertexID. Added a new #define
to the shader-parser to handle this, and updated the shader to use it.

Change-Id: Ia803e2efaf898cdbb93a63ef827cf34a01d8143a
Signed-off-by: David Steele <david.steele@samsung.com>