Fix anchor update issue when changing language
ClearStyleData clears not only font description runs but also color and underline runs.

Since the font client cache is cleared when the language is changed,
the font information in the text controller should be reset.

But style runs generated by the markup processor,
such as color and underline are not parsed, so should be maintained.

If the text is re-set due to a language change,
all style runs will be cleared and re-created, so there should be no issue.

Change-Id: I1da414fd8d9b9797d992ed32123020e4422d4b3c
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>