(SvgVisual) Request rasterize at SceneOn time if we use DesiredSize + Use ImageDimensions for last rasterized size
Let we request rasterization more faster if we are using DesiredSize.

Until now, we request rasterize at Relayout callback time, which is quite later than
other image cases ; SceneOn timing.

SVG need to sepcify the size of Visual. But if we are use DesiredSize, we don't need to
Relayout finished. So for this case, let we rasterize at SceneOn timing.

Also, clean up the last rasterized size store variable type as ImageDimensions, which is
ensurly integer type, so no numeric error occured.

Change-Id: Iacfcb126f5b8e631e7bec485b6755b0ea7a3a905
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>