Adjusting the stencil size in a text editor
In some cases, such as sw italic,
the glpyh rasterization result may be larger to the right than the actual advance.

Unlike a real italic font, this method cannot apply accurate kerning, hinting,
or advance width adjustments during the shaping stage,
which may result in glyphs extending beyond the right boundary and being clipped.

To work around adding padding to the right side of the text control,
increase the size of the stencil by the amount of right padding.

Change-Id: I27603ffdf467c67f8455a21f679aca318c06ebb4
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>