Set FocusableInTouch of NUI as true in WebView
- In NUI, FocusableInTouch defaults to false, but there was an issue where WebView would set this to true in Dali,
causing it to read as true even when the user hadn't set NUI's FocusableInTouch.
This code was previously removed to address that problem.
However, this caused IME (Input Method Editor) to not appear in multiple apps in DA that use NUI WebView.
Therefore, this code is being added back, and an explanation will be added to the NUI WebView constructor's comments.

Change-Id: I84ab4772e6e2a3710d89502be57d4c0b48e9867b